Replicate the same behavior without all the indirection
