+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/stream/drools/DroolsConstraintFactory.java
@@ -16,13 +16,6 @@
 
 package org.optaplanner.core.impl.score.stream.drools;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-import org.optaplanner.core.api.score.Score;
-import org.optaplanner.core.api.score.stream.Constraint;
 import org.optaplanner.core.api.score.stream.uni.UniConstraintStream;
 import org.optaplanner.core.impl.domain.constraintweight.descriptor.ConstraintConfigurationDescriptor;
 import org.optaplanner.core.impl.domain.solution.descriptor.SolutionDescriptor;
@@ -34,10 +27,8 @@ public final class DroolsConstraintFactory<Solution_> extends InnerConstraintFac
     private final SolutionDescriptor<Solution_> solutionDescriptor;
     private final String defaultConstraintPackage;
     private final DroolsVariableFactory variableFactory = new DroolsVariableFactory();
-    private final boolean droolsAlphaNetworkCompilationEnabled;
 
-    public DroolsConstraintFactory(SolutionDescriptor<Solution_> solutionDescriptor,
-            boolean droolsAlphaNetworkCompilationEnabled) {
+    public DroolsConstraintFactory(SolutionDescriptor<Solution_> solutionDescriptor) {
         this.solutionDescriptor = solutionDescriptor;
         ConstraintConfigurationDescriptor<Solution_> configurationDescriptor = solutionDescriptor
                 .getConstraintConfigurationDescriptor();
@@ -47,32 +38,12 @@ public final class DroolsConstraintFactory<Solution_> extends InnerConstraintFac
         } else {
             defaultConstraintPackage = configurationDescriptor.getConstraintPackage();
         }
-        this.droolsAlphaNetworkCompilationEnabled = droolsAlphaNetworkCompilationEnabled;
     }
 
     @Override
     public <A> UniConstraintStream<A> fromUnfiltered(Class<A> fromClass) {
         assertValidFromType(fromClass);
         return new DroolsFromUniConstraintStream<>(this, fromClass);
-    }
-
-    // ************************************************************************
-    // SessionFactory creation
-    // ************************************************************************
-
-    public <Score_ extends Score<Score_>> DroolsConstraintSessionFactory<Solution_, Score_> buildSessionFactory(
-            Constraint[] constraints) {
-        Map<String, List<Constraint>> constraintsPerIdMap = Arrays.stream(constraints)
-                .collect(Collectors.groupingBy(Constraint::getConstraintId));
-        constraintsPerIdMap.forEach((constraintId, constraintList) -> {
-            if (constraintList.size() > 1) {
-                throw new IllegalStateException(
-                        "There are multiple constraints with the same name in a package (" + constraintId + ").");
-            }
-        });
-
-        return new DroolsConstraintSessionFactory<>(solutionDescriptor, this, droolsAlphaNetworkCompilationEnabled,
-                constraints);
     }
 
     // ************************************************************************
